% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomData.R
\name{checkValiditycustomdata}
\alias{checkValiditycustomdata}
\title{Validity function for the CustomData class}
\usage{
checkValiditycustomdata(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{CustomData}}}
}
\value{
TRUE if the input data is valid, otherwise an error message
}
\description{
This function checks if the input data is a data frame
}
