% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomData.R
\name{colnames,CustomData-method}
\alias{colnames,CustomData-method}
\title{Method to get column names}
\usage{
\S4method{colnames}{CustomData}(x)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{CustomData}}}
}
\value{
Column names of the data frame contained in the CustomData object, or NULL if not applicable
}
\description{
This method returns the column names if the CustomData object contains a data frame
}
\examples{
df <- data.frame(a = 1:3, b = letters[1:3])
obj <- customdata(df)
colnames(obj)
}
\seealso{
\code{\link{customdata}}
}
