% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomData.R
\name{column_classes,CustomData-method}
\alias{column_classes,CustomData-method}
\title{Method to get column classes}
\usage{
\S4method{column_classes}{CustomData}(x)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{CustomData}}}
}
\value{
A named vector with the class of each column in the data frame
}
\description{
This method returns the class of each column in the data frame contained in the CustomData object
}
\examples{
df <- data.frame(a = 1:3, b = letters[1:3])
obj <- customdata(df)
column_classes(obj)
}
\seealso{
\code{\link{customdata}}
}
