% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomData.R
\docType{class}
\name{CustomData-class}
\alias{CustomData-class}
\title{An S4 class to represent CustomData}
\description{
An S4 class to represent CustomData
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame}
}}

