% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomData.R
\name{customdata}
\alias{customdata}
\title{Basic constructor of the CustomData class}
\usage{
customdata(data)
}
\arguments{
\item{data}{A data frame, vector, factor, or matrix to be used as the content of the CustomData object. All inputs will be converted to a data frame.}
}
\value{
An object of class \code{\linkS4class{CustomData}} with the specified data.
}
\description{
This function creates an object of class \code{\linkS4class{CustomData}}. The input data is converted to a data frame before being used to create the object.
}
\details{
Any input \code{data} will be internally converted to a data frame using \code{\link{as.data.frame}} before being assigned to the \code{data} slot of the \code{CustomData} object.
}
\examples{
# Create a CustomData object with different types of data
df <- data.frame(x = 1:3, y = letters[1:3])
obj_df <- customdata(df)
vec <- c(10, 20, 30)
obj_vec <- customdata(vec)
fac <- factor(c("A", "B", "A"))
obj_fac <- customdata(fac)
mat <- matrix(1:4, nrow = 2)
obj_mat <- customdata(mat)
}
\seealso{
\code{\link{checkValiditycustomdata}}
}
