% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize,CustomData-method}
\alias{normalize,CustomData-method}
\title{Method to normalize numeric columns of CustomData}
\usage{
\S4method{normalize}{CustomData}(x)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{CustomData}}.}
}
\value{
A new \code{\linkS4class{CustomData}} object with normalized numeric columns.
}
\description{
This method normalizes numeric columns in a \code{\linkS4class{CustomData}} object.
}
\details{
The normalization is applied only to numeric columns within the data frame of the \code{CustomData} object. Non-numeric columns remain unchanged. Numeric columns with only one unique value are left unchanged.
}
\examples{
# Create a CustomData object
df <- data.frame(a = 1:5, b = c(5, 5, 5, 5, 5), c = 10:14)
obj <- customdata(df)

# Normalize numeric columns
normalized_obj <- normalize(obj)
print(normalized_obj)
}
\seealso{
\code{\link{customdata}}
}
