% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_entropy}
\alias{plot_entropy}
\title{Function to visualize the normalized entropy for each factor column}
\usage{
plot_entropy(custom_data)
}
\arguments{
\item{custom_data}{A data frame where at least one column is factor}
}
\value{
A list containing a message confirming the plot generation, a vector of normalized entropies, and the names of the columns.
}
\description{
This function calculates the normalized entropy for each factor column in an object from CustomData and visualizes the results using a bar plot.
}
\details{
This function does not require extra packages
}
\examples{
df <- data.frame(a = factor(c(rep('a', 80), rep('b', 20))), 
                 b = factor(c(rep('x', 50), rep('y', 50))),
                 c = factor(c(rep('k', 70), rep('l', 30))))
obj <- customdata(df)                 
plot_entropy(obj)

}
