% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot ROC Curve from DataFrame or CustomData}
\usage{
plot_roc(data, prob_col, label_col)
}
\arguments{
\item{data}{An object of class \code{\linkS4class{CustomData}} or a data frame containing probability and binary label columns.}

\item{prob_col}{The name of the column in the data frame or CustomData object that contains probability values.}

\item{label_col}{The name of the column in the data frame or CustomData object that contains the binary labels.}
}
\value{
No return value; the function plots the ROC curve.
}
\description{
This function calculates ROC AUC and plots the ROC curve for a given data frame or CustomData object.
}
\examples{
df <- data.frame(prob = c(0.1, 0.4, 0.35, 0.8), label = c(0, 1, 0, 1))
plot_roc(df, "prob", "label")
}
\seealso{
\code{\link{rocAuc}} for extracting ROC points and calculate AUC, and \code{\link{plot_roc_curve}} for plotting the roc curve from the roc points
}
