% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{Plot ROC Curve}
\usage{
plot_roc_curve(roc_points, auc = NULL)
}
\arguments{
\item{roc_points}{A matrix where each row represents a point on the ROC curve, with columns for False Positive Rate (FPR) and True Positive Rate (TPR).}

\item{auc}{Optional; the Area Under the Curve (AUC) value to be displayed on the plot.}
}
\value{
No return value; the function plots the ROC curve.
}
\description{
This function plots the ROC curve given the ROC points and optionally displays the AUC value.
}
\examples{
df <- data.frame(prob = c(0.1, 0.4, 0.35, 0.8), label = c(0, 1, 0, 1))
result <- rocAuc(df, "prob", "label")
plot_roc_curve(result$roc_points, result$auc)
}
\seealso{
\code{\link{rocAuc}} for extracting ROC points and calculate AUC.
}
