% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCorrMatrix}
\alias{plotCorrMatrix}
\title{Plot Correlation Matrix using ggplot2}
\usage{
plotCorrMatrix(data)
}
\arguments{
\item{data}{Either a data frame or a CustomData object containing continuous variables.}
}
\value{
A ggplot2 object representing the correlation matrix.
}
\description{
This function plots a correlation matrix using ggplot2 for either a data frame or a CustomData object.
}
\details{
The function filters out non-numeric columns before calculating the correlation matrix. It calculates the correlation matrix and converts it to long format for plotting. The plot displays the correlation values with a color gradient and labels.
The function requires the ggplot2 and reshape2 packages to be installed.
}
\examples{
df <- data.frame(a = rnorm(100), b = rnorm(100), c = rnorm(100), d = as.factor(sample(letters, 100, replace = TRUE)))
plotCorrMatrix(df)

# Example with CustomData
custom_data <- customdata(df)
plotCorrMatrix(custom_data)
}
