% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomData.R
\name{show,CustomData-method}
\alias{show,CustomData-method}
\title{Print method for CustomData}
\usage{
\S4method{show}{CustomData}(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{CustomData}}}
}
\value{
Prints details of the CustomData object, including its type and content
}
\description{
This method prints information about the CustomData object
}
\examples{
df <- data.frame(a = 1:5, b = letters[1:5])
obj <- customdata(df)
show(obj)
vec <- c(1, 2, 3)
obj_vec <- customdata(vec)
show(obj_vec)
}
\seealso{
\code{\link{customdata}}
}
